requirements.txt
fastapi==0.112.0
uvicorn==0.22.0	
pydantic
python-dotenv==1.0.1
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.5
langchain-chroma
langchain-objectbox==0.1.0
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langchainhub==0.1.20
langdetect==1.0.9
langserve==0.2.2
langsmith==0.1.98
PyMuPDF==1.24.9
chromadb==0.5.3
chroma-hnswlib==0.7.3
PyYAML==6.0.2
pysqlite3-binary
requests
chardet